openpgp: Drop redundant recursive_depth field.
  - The packet parser's recursive depth can be computed from the path,
    which we now track.  As such, don't track the recursive depth
    separately, just derive it from the path.

+++ ffi/src/openpgp.rs
@@ -1241,7 +1241,7 @@ pub extern "system" fn sq_packet_parser_recursion_depth
     (pp: Option<&PacketParser>)
      -> uint8_t {
     let pp = pp.expect("PacketParser is NULL");
-    pp.recursion_depth
+    pp.recursion_depth() as u8
 }
 
 /// Finishes parsing the current packet and starts parsing the
